requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
chromadb
openai
fastapi
uvicorn
watchdog
python-dotenv
unstructured
